Refactor layout, created book item, and styles e colors.
