+++ Trilha Android/Aula 03/.idea/misc.xml
@@ -3,7 +3,9 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Learning/ioasysCamp/Trilha Android/Aula 03/app/src/main/res/layout/activity_book_list.xml" value="0.3546875" />
         <entry key="..\:/Learning/ioasysCamp/Trilha Android/Aula 03/app/src/main/res/layout/activity_login.xml" value="0.33425925925925926" />
+        <entry key="..\:/Learning/ioasysCamp/Trilha Android/Aula 03/app/src/main/res/layout/book_item.xml" value="0.3546875" />
         <entry key="app/src/main/res/drawable/bg_input.xml" value="0.35520833333333335" />
         <entry key="app/src/main/res/layout/activity_login.xml" value="0.25" />
       </map>
