+++ Trilha Android/Aula 03/app/src/main/java/br/com/pedroabreudev/books/activity/LoginActivity.kt
@@ -1,10 +1,10 @@
-package br.com.pedroabreudev.books
+package br.com.pedroabreudev.books.activity
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatTextView
+import br.com.pedroabreudev.books.R
 import com.google.android.material.button.MaterialButton
 
 class LoginActivity : AppCompatActivity() {
