+++ Trilha Android/Aula 03/app/src/main/res/layout/activity_book_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_book_list"
-    tools:context=".BookListActivity">
+    tools:context=".activity.BookListActivity">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/txtLabel"
@@ -54,6 +54,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/edSearch" />
+        app:layout_constraintTop_toBottomOf="@id/edSearch"
+        tools:listitem="@layout/book_item" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
