+++ Trilha Android/Aula 03/app/src/main/res/layout/book_item.xml
@@ -1,6 +1,77 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="wrap_content"
+    android:layout_marginHorizontal="16dp"
+    android:layout_marginVertical="4dp"
+    android:background="@android:color/white"
+    android:elevation="20dp"
+    app:cardCornerRadius="4dp">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="16dp"
+        android:paddingVertical="28dp">
+
+        <androidx.appcompat.widget.AppCompatImageView
+            android:id="@+id/imgBook"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/img_book"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/tvBookTitle"
+            style="@style/Book_Title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintLeft_toRightOf="@+id/imgBook"
+            app:layout_constraintTop_toTopOf="@+id/imgBook"
+            tools:text="Creative Selection" />
+
+        <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/tvBookSubTitle"
+            style="@style/Book_SubTitle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintLeft_toRightOf="@+id/imgBook"
+            app:layout_constraintTop_toBottomOf="@id/tvBookTitle"
+            tools:text="Ken Kocienda" />
+
+        <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/tvBookPages"
+            style="@style/Book_Description"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="4dp"
+            app:layout_constraintBottom_toTopOf="@id/tvBookEditor"
+            app:layout_constraintLeft_toRightOf="@id/imgBook"
+            tools:text="150 Páginas" />
+
+        <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/tvBookEditor"
+            style="@style/Book_Description"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="4dp"
+            app:layout_constraintBottom_toTopOf="@id/tvBookDate"
+            app:layout_constraintLeft_toRightOf="@id/imgBook"
+            tools:text="Editora Loyola" />
+
+        <androidx.appcompat.widget.AppCompatTextView
+            android:id="@+id/tvBookDate"
+            style="@style/Book_Description"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintBottom_toBottomOf="@id/imgBook"
+            app:layout_constraintLeft_toRightOf="@id/imgBook"
+            tools:text="Publicado em 2020" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</com.google.android.material.card.MaterialCardView>
