+++ Trilha Android/Aula 03/app/src/main/res/values/colors.xml
@@ -9,4 +9,10 @@
     <color name="white">#FFFFFFFF</color>
     <color name="pink">#B22E6F</color>
     <color name="background_textInputLayout">#52000000</color>
+
+    <color name="title">#333333</color>
+    <color name="title_50">#80333333</color>
+    <color name="subtitle">#AB2680</color>
+    <color name="description">#999999</color>
+
 </resources>
